+++ Kernel/Modules/CustomerTicketProcess.pm
@@ -223,8 +223,20 @@ sub Run {
         }
     }
 
+    # list only Active processes by default
+    my @ProcessStates = ('Active');
+
+    # set AJAXDialog for proper error responses, screen display and process list
+    $Self->{AJAXDialog} = $Self->{ParamObject}->GetParam( Param => 'AJAXDialog' ) || '';
+
+    # fetch also FadeAway processes to continue working with existing tickets, but not to start new
+    #    ones
+    if ( !$Self->{AJAXDialog} && $Self->{Subaction} ) {
+        push @ProcessStates, 'FadeAway'
+    }
+
     my $ProcessList = $Self->{ProcessObject}->ProcessList(
-        ProcessState => ['Active'],
+        ProcessState => \@ProcessStates,
         Interface    => ['CustomerInterface'],
     );
 
@@ -267,6 +279,29 @@ sub Run {
             %Param,
             ProcessList     => $ProcessList,
             ProcessEntityID => $ProcessEntityID
+        );
+    }
+
+    # check if the selected process from the list is valid, prevent tamper with process selection
+    #    list (not existing, invalid an fade away processes must not be able to start a new process
+    #    ticket)
+    elsif (
+        $Self->{Subaction} eq 'DisplayActivityDialogAJAX'
+        && !$ProcessList->{$ProcessEntityID}
+        && $Self->{AJAXDialog}
+        )
+    {
+
+        # translate the error message (as it will be injected in the HTML)
+        my $ErrorMessage
+            = $Self->{LayoutObject}->{LanguageObject}->Get("The selected process is invalid!");
+
+        # return a predefined HTML sctructure as the AJAX call is expecting and HTML response
+        return $Self->{LayoutObject}->Attachment(
+            ContentType => 'text/html; charset=' . $Self->{LayoutObject}->{Charset},
+            Content     => '<div class="ServerError" data-message="' . $ErrorMessage . '"></div>',
+            Type        => 'inline',
+            NoCache     => 1,
         );
     }
 
@@ -739,7 +774,7 @@ sub _GetParam {
         $ActivityEntityID = $Ticket{
             'DynamicField_'
                 . $Self->{ConfigObject}->Get("Process::DynamicFieldProcessManagementActivityID")
-            };
+        };
         if ( !$ActivityEntityID ) {
             $Self->{LayoutObject}->CustomerFatalError(
                 Message =>
@@ -3274,7 +3309,7 @@ sub _StoreActivityDialog {
         $ActivityEntityID = $Ticket{
             'DynamicField_'
                 . $Self->{ConfigObject}->Get('Process::DynamicFieldProcessManagementActivityID')
-            };
+        };
         if ( !$ActivityEntityID )
         {
             return $Self->{LayoutObject}->CustomerErrorScreen(
@@ -3286,7 +3321,7 @@ sub _StoreActivityDialog {
         $ProcessEntityID = $Ticket{
             'DynamicField_'
                 . $Self->{ConfigObject}->Get('Process::DynamicFieldProcessManagementProcessID')
-            };
+        };
 
         if ( !$ProcessEntityID )
         {