+++ CHANGES.md
@@ -1,4 +1,5 @@
 #3.3.0 RC1 ????-??-??
+ - 2013-10-28 Fixed bug#[9897](http://bugs.otrs.org/show_bug.cgi?id=9897) - process ticket 'fading away' -> Process is invalid.
  - 2013-10-28 Fixed bug#[9850](http://bugs.otrs.org/show_bug.cgi?id=9850) - Dashboard Filter error in ticket and page count.
  - 2013-10-28 Fixed bug#[9888](http://bugs.otrs.org/show_bug.cgi?id=9888) - Warning when using standard template.
  - 2013-10-28 Fixed bug#[9749](http://bugs.otrs.org/show_bug.cgi?id=9749) - Using spell checker in rich text Editor strips HTML formatting.